ISIS-972 Make it possible to set FontAwesome icons after the label for action links
Make it possible to set FontAwesome icons after the label for css menu items
