+++ component/viewer/wicket/ui/src/main/java/org/apache/isis/viewer/wicket/ui/components/actionmenu/serviceactions/CssSubMenuItemsPanel.html
@@ -23,7 +23,6 @@
     <ul class="cssSubMenuItemsPanel">
         <li wicket:id="subMenuItems" class="cssSubMenuItem">
             <a wicket:id="menuLink" class="btn btn-primary btn-sm">
-                <span class="fontAwesomeIcon" wicket:id="menuLinkFontAwesome"></span>
                 <span class="menuLabel" wicket:id="menuLabel">[menu label]</span>
             </a>
             <p wicket:id="menuLabel">[menu label]</p>
